+++ product_service/src/utils/headers.ts
@@ -1,0 +1,5 @@
+export const commonHeaders = {
+  "Access-Control-Allow-Methods": "OPTIONS, GET, POST, PUT",
+  "Access-Control-Allow-Origin": "*",
+  "Content-Type": "application/json",
+};
